linux version

+++ include/raptor/types.h
@@ -23,18 +23,19 @@
 extern "C" {
 #endif
 
+#include <stddef.h>
 #include <stdint.h>
 
 typedef uint64_t ConnectionId;
 typedef uint64_t raptor_connection_t;
 
 typedef struct {
-    uint16_t max_connections;
-    uint16_t send_recv_timeout;
-    uint16_t connection_timeout;
-    uint16_t max_package_per_second;
-    uint16_t send_recv_threads;
-    uint16_t accept_threads;
+    size_t max_connections;
+    size_t send_recv_timeout;
+    size_t connection_timeout;
+    size_t max_package_per_second;
+    size_t send_recv_threads;
+    size_t accept_threads;
 } raptor_options_t;
 
 typedef raptor_options_t RaptorOptions;
